chore: apply review comments
