+++ tests-fuzz/src/translator.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod common;
 pub mod mysql;
 pub mod postgres;
